Update Dear_Will.md
at line 6 saying you aren't very rational feels very harsh in this context; i might approach it from the other side and say you're romanticizing it, or something of the sort

similarly, in lines 9-12, your concerns are voiced very concretely and you use the word productive twice. it's a bit of a dense paragraph; is there a way you could rephrase this to express the idea of productivity without mentioning it twice?

the paragraph at lines 29-34 could be condensed, too. it's a long apology and meandering could work but it feels as if the same thing is being said in a variety of ways as it stands now...

i made some smaller changes without commenting (please view them as suggestions) to eliminate the wordiness... 

may the day always stoop to greet you, my chap!