Update Postmortem:_Corrupting_RSS_And_Twitter.md
it might be worthwhile to restructure the first two sentences of the post as they both follow the same flow pattern.

lines 14-16 are also a bit jumbled; can you separate those sentences somehow to improve clarity?

the sentence at line 19 is good, but it might be wiser to put it after the explanation of how the twitter feed is generated.

likewise, the timing in lines 33-34 is off, try "Then I decided to write a quick blog post advertising the change; in under an hour, I had written, edited, and published the post."

in line 36 "it took a little to respond": are you referring to the amount of time it took? or emotional coaxing? in either case, i would be more specific about what it needed