Update zepto,_An_Eulogy.md
vor weg: i sniffed around a bit and am now fairly certain that "a eulogy" is more correct than "an eulogy", even if it does not abide by the rules of writing (consonant and vowel word beginnings), it does abide by the rules of phonology, and writing has decided to try to be in agreement with that.
the paragraph that starts at line 24 is really strong!

you may want to consider swapping the word "simple" out for something else in a few spots.

a beautiful farewell, turtle. well done!