Update Reading,_The_First_Year.md
In line 20, it might be a good idea to change the font stylization on "technical papers" and "books" to indicate that they're categories.

The paragraph starting at 28 reads a little harsh. Of course it's your decision to include it, but it might be a bit demotivating for your less productive readers, despite the nice turn around in the following paragraph.

otherwise, hurrah!